Avoid incorrect copyright claims
When building this in 2025, it claims Copyright 2025
without any copyright-worthy work happening in that year.

Also note that the number does not have to be updated every year
as copyright does not expire for 70+ years.

This patch was done while working on reproducible builds for openSUSE, sponsored by the NLnet NGI0 fund.

Signed-off-by: Bernhard M. Wiedemann <bwiedemann@suse.de>