COMMON: Enhance size checking for RSA-OAEP encryption/wrapping
Only a certain amount of data can be encrypted with RSA-OAEP for a given
RSA key size and OAEP digest. The data to encrypt can only be up to
'modulus_bytes - 2 * hashlen - 2' bytes. Check this already for a
length-only call and return CKR_DATA_LEN_RANGE if the amount of data
can not be encrypted.

For wrapping with CKM_RSA_AES_KEY_WRAP, return CKR_WRAPPING_KEY_SIZE_RANGE
in case the temporary AES key can not be wrapped with RSA-OAEP with the
used RSA key size and OAEP digest.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>