ICA: Report mechanisms dependent on if libica is in FIPS mode
Add mechanisms where we have a SW fallback unconditionally, but only
if libica is not in FIPS mode. If libica is in FIPS mode, the loop in
over the mechanisms supported by libica will add those mechanisms that
libica supports and that are FIPS approved. Mechanisms that are not
FIPS approved will not be added, even if there is a SW fallback.

Because the SW fallback is OpenSSL-based, and OpenSSL is most likely
also in FIPS mode in this case, the SW fallback would also fail if using
an algorithm that is not FIPS approved.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>