POLICY: Fix signature size check for the Dilithium mechanism
Handling for mechanism CKM_IBM_DILITHIUM is missing in function
policy_check_signature_size(). This causes policy_update_mech_info() to
remove the CKF_SIGN and CKF_VERIFY flags when a minimum strength of
larger than 0 is configured in the policy config file, because function
policy_check_signature_size() erroneously returns a signature size of zero,
which is below the allowed signature size. Return a signature size of
256 bits instead.

This only affects the mechanism info returned by C_GetMechanismInfo(),
when the policy defines minimum strength of larger than 0. It does not
affect the mechanism execution, nor signature creation or verification.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>