PKEY: Fix out-of-bounds access in Edwards curve verify
When PKEY is used for an Edwards curve verify, then the public key is copied
into the CPACF parameter block. Add a length check for the public key size,
and also make sure to only copy up to the size of the public key for Ed448.
The public key field in the CPACF parameter block for Ed448 is 64 bytes in
size, while the public key of Ed448 is only 57 bytes. Thus, only copy
57 bytes, otherwise the public key buffer would be accessed outside of its
size. For Ed25519 the parameter block field has the same size as the public
key (i.e. 32 bytes), but better check the public key size, too.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>