EP11: Check session-bound keys when CKA_TOKEN or CKA_PRIVATE is changed
In case attributes CKA_TOKEN and/or CKA_PRIVATE are changed (only allowed
during C_CopyObject, but not via C_SetAttributeValues), special checking is
required in regards of session bound keys. In case strict session mode,
VHSM mode or FIPS session mode is active, the changed attributes may cause
the key object to be expected to be bound to a different EP11 session as
it currently is. Check if the key object is bound to the expected
session, and fail if not. This prevents to have inconsistent sessions of
key objects.

Note: As of now, EP11 does not support/allow to change key objects in
regards of the session they are bound to, so the only option currently is
to fail.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>