COMMON/SOFT/ICA: Support key wrapping with AES-GCM
As per PKCS#11 standard, CKM_AES_GCM is supposed to support key wrapping.
The Soft and ICA tokens both support CKM_AES_GCM, but do not support key
wrapping with it so far.

Update the common code to support key wrapping with CKM_AES_GCM and flag
the mechanism accordingly in the Soft and ICA tokens.

Closes: https://github.com/opencryptoki/opencryptoki/issues/848

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>