COMMON: Fix error with C_WrapKey size-query calls
When C_WrapKey is called with parameter pWrappedKey being NULL, then this is
a so called size-query call, to obtain the required size of the output buffer.

In such cases, the to-be-encrypted key material must still be obtained to
be able to determine the size of the encrypted key material.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>