EP11: Add support for CKM_IBM_ML_KEM_WITH_ECDH mechanism
Add support for mechanism CKM_IBM_ML_KEM_WITH_ECDH for the EP11 token.

The CKM_IBM_ML_KEM_WITH_ECDH mechanism is only supported if also
mechanisms CKM_IBM_ML_KEM and CKM_ECDH1_DERIVE are supported, i.e.
when all APQNs configure for the token have the required firmware level
(9.5 or later on IBM z17, and 8.37 or later on IBM z16) to support them,
and when the EP11 host library version is 4.2 or later.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>