PKCSSLOTD: Check for duplicate tokens with no 'tokname'
Having multiple identical tokens in opencryptoki.conf can cause various
problems, because the token directory containing the token specific data
as well as token objects is then shared between the duplicate tokens.

When the token configuration contains a 'tokname', then there is already
a check that no duplicate token name is used. However, for tokens without
a 'tokname' pkcsslotd is not able to check for duplicates, because only
the token DLL knows the hard-coded token name.

Use the token's DLL name to check for duplicates in such cases. Two tokens
that use the same DLL name will use the same hard coded token name.

Of course, one can build a situation where different DLL names still point
to the exact shared object (e.g. using links), and this would erroneously be
treated as different tokens, although it would be the same token. There is
no easy way to take care for such tricks, so lets ignore this for now.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>