COMMON: Allow filter of ock_generic_get_mechanism_info() to adjust mech info
When the filter function of the ock_generic_get_mechanism_info() and
ock_generic_get_mechanism_list() functions is called with a non-NULL info
parameter, then the filter function can adjust the mechanism info to be
returned to the application. This, for example, allows to modify the
supported key sizes of a mechanism dependent on the token's state.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>