COMMON/SOFT/ICA/EP11: Add support for SHA512-224/256 key derivation mechanisms
Add support for the mechanism CKM_SHA512_224_KEY_DERIVATION and
CKM_SHA512_256_KEY_DERIVATION to common code, as well as to the Soft, ICA,
and EP11 token.

Not that the EP11 firmware does currently not support these mechanisms, so
the EP11 token won't support them either. However, if the EP11 firmware ever
supports them, then the EP11 token is ready for it.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>