COMMON: Fix AES-CTR counter width range checking
The range check for the AES-CTR counter width is off by a factor of 8.

According to the PKCS#11 standard, counter width is expressed in bits, and
the range is supposed to be: 0 < ulCounterBits ≤ 128.

This affects the soft token, as well as the ICA token if the software
fallback path is taken.

Closes: https://github.com/opencryptoki/opencryptoki/issues/881

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>