COMMON: Fix AES-CTR counter width bits vs bytes confusion
According to the PKCS#11 standard, counter width is expressed in bits.
When the counter width is used as length for a memcpy, it must be divided
by 8 to get the width in bytes.

Note that common code functions ckm_aes_ctr_encrypt() and
ckm_aes_ctr_decrypt() ensure that the counter width is a multiple of
8 bits, so it can be safely divided by 8.

This affects the soft token, as well as the ICA token if the software
fallback path is taken.

Closes: https://github.com/opencryptoki/opencryptoki/issues/883

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>