COMMON: Do not leave incomplete files behind if process is killed
When token files like NVTOK.DAT, MK_SO, MK_USER, OBJ.IDX, or token object
files are created or updated, and the process is killed, then it may happen
that not all file contents has been flushed to disk at the time the process
gets killed. This may leave incomplete or zero size files behind, that may
cause to token to be no longer usable (e.g. if NVTOK.DAT is corrupt).

Update such files in an atomic way: First write the new file contents into a
separate temporary file (e.g. using a .TMP postfix), and then rename that
temporary file to the original name. Note that the process holds the token
lock while updating or creating token files, so no other process can use the
file in question concurrently.

From the rename() man page:

       If newpath already exists, it will be atomically replaced, so that
       there is no point at which another process attempting to access
       newpath will find it missing.

That way, if the process gets killed, either the old file is still there,
or it was replaced by the new file. In either case, the file is complete
and consistent. In worst case the temporary file is left behind, and the
update is lost.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>