CCA: Trim RSA public exponent when retrieved from CCA key tokens
During RSA key generation as well as during RSA key blob import, the public
key attributes CKA_MODULUS and CKA_PUBLIC_EXPONENT are populated by retrieving
the respective values from the private or public secure key token section.

The key components inside a CCA private or public secure key token section
might be padded with binary zeros on the left, and the length information
contained in the key token includes this padding. This is especially the
case for the public exponent.

Because the RSA key components are big endian integers, leading zeros do not
change the numerical value. However attributes such as CKA_PUBLIC_EXPONENT
should contain the numerical value without any leading zeros.

Trim the public exponent value to remove any leading zeros for the public
exponent value. The modulus value and length is left as is, because the
modulus length is always exactly the bit size of the RSA key. Even if the
modulus value would ever contain leading zeros, the modulus length should
be kept as is.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>