P11SAK: Add support for EC curve BLS12-381
Make BLS12-381 known to p11sak, so that it can generate EC keys of this curve.

Public EC key objects of the BLS12-381 curve needs to have CKA_DERIVE=TRUE to
be used with the aggregate mechanism. Thus, allow to specify the CKA_DERIVE
attribute also for the public part of an EC key.

Note: As per PKCS#11 standard, public keys don't have the CKA_DERIVE attribute,
but we allow it anyway to be able to support BLS12-384 aggregation operations.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>